Minor tweaks to reduce build time marginally
These minor improvements will not shorten build time drastically, but
that's better than nothing

1. Increase items per page from 5 to 10 in pagination

This reduces the number of paginator pages being created in each build, which
also enhance user experience, because it seems to me that a paginator page with
just 5 items on it is too short.

2. Reorganise tags

This minimises the possibility of having tags with just one or two items
associated with it, which makes those tags useless.